requirement.txt
aiofiles==23.2.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
appdirs==1.4.4
attrs==23.2.0
Brotli @ file:///D:/bld/brotli-split_1695989908365/work
certifi @ file:///C:/b/abs_35d7n66oz9/croot/certifi_1707229248467/work/certifi
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1698833585322/work
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
docker-pycreds==0.4.0
einops==0.7.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
fastapi==0.109.0
ffmpy==0.3.1
filelock @ file:///C:/b/abs_f2gie28u58/croot/filelock_1700591233643/work
fonttools==4.47.2
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.41
gmpy2 @ file:///C:/ci/gmpy2_1645438895476/work
gradio==4.15.0
gradio_client==0.8.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.3
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1701026962277/work
importlib-resources==6.1.1
Jinja2 @ file:///C:/b/abs_f7x5a8op2h/croot/jinja2_1706733672594/work
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
largestinteriorrectangle==0.2.0
llvmlite==0.42.0
markdown-it-py==3.0.0
MarkupSafe @ file:///C:/b/abs_ecfdqh67b_/croot/markupsafe_1704206030535/work
matplotlib==3.8.2
mdurl==0.1.2
MouseInfo==0.1.3
mpmath @ file:///C:/b/abs_7833jrbiox/croot/mpmath_1690848321154/work
networkx @ file:///C:/b/abs_e6gi1go5op/croot/networkx_1690562046966/work
numba==0.59.1
numpy==1.26.4
opencv-python==4.6.0
openpyxl==3.1.2
orjson==3.9.12
packaging==23.2
pandas==2.2.1
pathtools==0.1.2
pillow==10.3.0
protobuf==4.25.2
psutil==5.9.8
PyAutoGUI==0.9.54
pydantic==2.5.3
pydantic_core==2.14.6
pydub==0.25.1
PyGetWindow==0.0.9
Pygments==2.17.2
PyMsgBox==1.0.9
pyparsing==3.1.1
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
PySocks @ file:///D:/bld/pysocks_1661604991356/work
python-dateutil==2.8.2
python-multipart==0.0.6
pytweening==1.2.0
pytz==2023.3.post1
PyWavelets==1.6.0
PyYAML @ file:///C:/b/abs_782o3mbw7z/croot/pyyaml_1698096085010/work
referencing==0.32.1
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1684774241324/work
rich==13.7.0
rpds-py==0.17.1
ruff==0.1.14
scipy==1.13.0
semantic-version==2.10.0
sentry-sdk==1.39.2
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
starlette==0.35.1
stitching==0.5.3
sympy @ file:///C:/b/abs_82njkonm7f/croot/sympy_1701397685028/work
timm==0.6.12
tomlkit==0.12.0
toolz==0.12.1
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tqdm==4.64.1
trimesh==3.9.42
typer==0.9.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1702176139754/work
tzdata==2023.4
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1699933488691/work
uvicorn==0.27.0
wandb==0.13.9
websockets==11.0.3
win-inet-pton @ file:///D:/bld/win_inet_pton_1667051142467/work
